sandbox/wolf/requirements.txt
flask
flask-socketio
eventlet
docker-compose